requirements.txt
Flask
Flask-Mail
Celery
redis
python-dotenv
gunicorn
